requirements.txt
asgiref==3.2.3
astroid==2.3.3
autopep8==1.5
cement==2.8.2
certifi==2019.11.28
chardet==3.0.4
colorama==0.3.9
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.3
django-stdimage==5.0.2
docutils==0.15.2
future==0.16.0
gunicorn==20.0.4
idna==2.7
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
pathspec==0.5.9
Pillow==7.0.0
progressbar2==3.47.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pylint==2.4.4
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.3
PyYAML==5.2
requests==2.20.1
semantic-version==2.5.0
six==1.11.0
sqlparse==0.3.1
static3==0.7.0
termcolor==1.1.0
typed-ast==1.4.1
urllib3==1.24.3
wcwidth==0.1.8
wrapt==1.11.2
